feat(orders): add order detail page with address, items, and price breakdown
Adds a GET /api/orders/:id endpoint returning order status, items (with
product image/category), shipping cost, and decrypted delivery address.
The new /orders/:id customer page shows the delivery timeline, itemized
products with images, the delivery address, and a subtotal/shipping/total
summary, linked from "View Details" on each order card.

Order items across both the list and detail views now render the actual
product image (via product category + name) instead of a placeholder
letter, falling back to the letter avatar if no image is found.

Shared order list helpers/components are extracted from OrdersPage into
orderHelpers.js and orderComponents.jsx for reuse on the new detail page.
